requirements.txt
Django~=5.0.4
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pyjwt
pytz
sqlparse
psycopg2-binary
python-dotenv